1.0.6 - base message fix
