+++ src/index.ts
@@ -1,3 +1,4 @@
+export * as API from "revolt-api";
 export * from "./api";
 export * from "./Client";
 export * from "./managers";
